requirements.txt
streamlit
pandas
numpy
scikit-learn
transformers
torch
PyMuPDF
PyPDF2
fpdf
python-dotenv
openai>=1.0.0
pydrive2
seaborn
matplotlib
openpyxl
pandas
numpy
scikit-learn
python-dotenv
langchain
huggingface_hub
ipywidgets
openpyxl
langchain
langchain-community
streamlit
plotly
seaborn
matplotlib
streamlit_option_menu
streamlit-extras
numerize
altair
datasets